service-workers/service-worker/next-hop-protocol.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Service Worker: Verify nextHopProtocol is set correctly</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>

async function getNextHopProtocol(frame, url_suffix) {
  let url = new URL(`resources/empty.js${url_suffix}`, self.location).href;
  await frame.contentWindow.fetch(url).then(r => r.text());
  let entryList = frame.contentWindow.performance.getEntriesByName(url);
  let entry = entryList[entryList.length - 1];
  return entry.nextHopProtocol;
}

promise_test(async (t) => {
  const scope = 'resources/empty.html?next-hop-protocol';
  const script = 'resources/fetch-rewrite-worker.js';
  const expected_h1_protocol = 'http/1.1';
  let frame;

  const registration =
      await service_worker_unregister_and_register(t, script, scope);
  t.add_cleanup(async _ => registration.unregister());
  await wait_for_state(t, registration.installing, 'activated');
  frame = await with_iframe(scope);
  t.add_cleanup(_ => frame.remove());

  assert_equals(await getNextHopProtocol(frame, '?generate-png'), '',
                'nextHopProtocol is not set on synthetic response');
  assert_equals(await getNextHopProtocol(frame, '?ignore'), 'http/1.1',
                'nextHopProtocol is set on fallback');
  assert_equals(await getNextHopProtocol(frame, ''), 'http/1.1',
                'nextHopProtocol is set on pass-through');
  assert_equals(await getNextHopProtocol(frame, '?cache'), 'http/1.1',
                'nextHopProtocol is set on cached response');
}, 'nextHopProtocol is reported correctly when routed via a service worker.');

</script>
